+++ Frends.Files.LocalBackup/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [2.1.0] - 2023-05-31
+### Added
+- Added parameter for FilePaths. If FilePaths is used it will overwrite the SourceDirectory and SourceFile parameters.
+
+## [2.0.4] - 2023-05-09
+### Fixed
+- Fixed issue where Cleanup adds deleted message even if files were not deleted.
+
 ## [2.0.3] - 2022-11-25
 ### Fixed
 - Fixed issue when using CreateSubdirectories Cleanup uses same directory which is created in Backup and when there's no files to backup the directory is deleted and cleanup won't find the directory.